Several fixes especially for windows: alWxFileList: setting the report style in the constructor + fixed bug in image viewer + managed to compile 64 bits with working wrapAlgorithm allow 2 external wrapped libraries + fixed bug with size_t previous changes
git-svn-id: file:///home/karl/SVN2/amilab.svn/code/trunk@1726 52b0d089-85ac-4fb5-acf5-4e3172ea7aaf

+++ src/LanguageBase/paramlist.h
@@ -70,7 +70,7 @@ private:
    * 
    * @return the number of parameters
    */
-  size_t GetNumParam() { return params.size(); }
+  int GetNumParam() { return (int)params.size(); }
 
   /**
    * 
@@ -116,7 +116,7 @@ class ParamListDecl {
     {
     }
 
-  size_t GetNumParam() { return num_param; }
+  int GetNumParam() { return num_param; }
 
   vartype GetType( int i) { 
     if (i<0) {
